sqlite: reference multipart uploads ID and get rid of HasMultipartUpload

+++ sia/persist/sqlite/multipart.go
@@ -10,15 +10,6 @@ import (
 	"lukechampine.com/frand"
 )
 
-func parseUploadID(s string) (uid [16]byte, _ error) {
-	if len(s) != 32 {
-		return uid, fmt.Errorf("invalid length: got %d, want 32", len(s))
-	} else if _, err := hex.Decode(uid[:], []byte(s)); err != nil {
-		return uid, fmt.Errorf("failed to parse upload ID %q: %w", s, err)
-	}
-	return
-}
-
 // CreateMultipartUpload persists metadata for a new multipart upload and
 // returns a random upload ID.
 func (s *Store) CreateMultipartUpload(bucket, name string, meta map[string]string) (string, error) {
@@ -53,20 +44,18 @@ func (s *Store) CreateMultipartUpload(bucket, name string, meta map[string]strin
 
 // AbortMultipartUpload removes a multipart upload from the store.
 func (s *Store) AbortMultipartUpload(bucket, name, uploadID string) error {
-	uid, err := parseUploadID(uploadID)
-	if err != nil {
-		return err
-	}
 	return s.transaction(func(tx *txn) error {
 		bid, err := bucketID(tx, bucket)
 		if err != nil {
 			return err
 		}
 
-		res, err := tx.Exec(`
-			DELETE FROM multipart_uploads
-			WHERE bucket_id = $1 AND name = $2 AND upload_id = $3
-		`, bid, name, sqlUploadID(uid))
+		uid, err := multipartID(tx, uploadID, bid, name)
+		if err != nil {
+			return err
+		}
+
+		res, err := tx.Exec(`DELETE FROM multipart_uploads WHERE id = $1`, uid)
 		if err != nil {
 			return err
 		} else if n, err := res.RowsAffected(); err != nil {
@@ -79,47 +68,28 @@ func (s *Store) AbortMultipartUpload(bucket, name, uploadID string) error {
 	})
 }
 
-// HasMultipartUpload returns an error if the multipart upload does not exist.
-func (s *Store) HasMultipartUpload(bucket, name, uploadID string) error {
-	uid, err := parseUploadID(uploadID)
-	if err != nil {
-		return err
-	}
+// AddMultipartPart adds metadata for a multipart part to the store.
+func (s *Store) AddMultipartPart(bucket, name, uploadID string, partNumber int) error {
 	return s.transaction(func(tx *txn) error {
 		bid, err := bucketID(tx, bucket)
 		if err != nil {
 			return err
 		}
-		var exists bool
-		if err := tx.QueryRow(`
-			SELECT EXISTS(
-				SELECT 1 FROM multipart_uploads
-				WHERE bucket_id = $1 AND name = $2 AND upload_id = $3
-			)
-		`, bid, name, sqlUploadID(uid)).Scan(&exists); err != nil {
-			return err
-		} else if !exists {
-			return s3errs.ErrNoSuchUpload
-		}
-		return nil
-	})
-}
 
-// AddMultipartPart adds metadata for a multipart part to the store.
-func (s *Store) AddMultipartPart(uploadID string, partNumber int) error {
-	uid, err := parseUploadID(uploadID)
-	if err != nil {
-		return err
-	}
-	return s.transaction(func(tx *txn) error {
+		uid, err := multipartID(tx, uploadID, bid, name)
+		if err != nil {
+			return err
+		}
+
 		res, err := tx.Exec(`
-			INSERT INTO multipart_parts (upload_id, part_number, created_at)
+			INSERT INTO multipart_parts (multipart_upload_id, part_number, created_at)
 			VALUES ($1, $2, $3)
-			ON CONFLICT(upload_id, part_number) DO UPDATE SET created_at = EXCLUDED.created_at
-		`, sqlUploadID(uid), partNumber, sqlTime(time.Now()))
+			ON CONFLICT(multipart_upload_id, part_number) DO UPDATE SET created_at = EXCLUDED.created_at
+		`, uid, partNumber, sqlTime(time.Now()))
 		if err != nil {
 			return fmt.Errorf("failed to insert multipart part: %w", err)
 		}
+
 		n, err := res.RowsAffected()
 		if err != nil {
 			return err
@@ -131,12 +101,18 @@ func (s *Store) AddMultipartPart(uploadID string, partNumber int) error {
 }
 
 // FinishMultipartPart updates metadata for a multipart part in the store.
-func (s *Store) FinishMultipartPart(uploadID string, partNumber int, contentMD5 [16]byte, contentSHA256 *[32]byte, contentLength int64) error {
-	uid, err := parseUploadID(uploadID)
-	if err != nil {
-		return err
-	}
+func (s *Store) FinishMultipartPart(bucket, name, uploadID string, partNumber int, contentMD5 [16]byte, contentSHA256 *[32]byte, contentLength int64) error {
 	return s.transaction(func(tx *txn) error {
+		bid, err := bucketID(tx, bucket)
+		if err != nil {
+			return err
+		}
+
+		uid, err := multipartID(tx, uploadID, bid, name)
+		if err != nil {
+			return err
+		}
+
 		var sha256Bytes []byte
 		if contentSHA256 != nil {
 			sha256Bytes = contentSHA256[:]
@@ -144,16 +120,17 @@ func (s *Store) FinishMultipartPart(uploadID string, partNumber int, contentMD5 
 		res, err := tx.Exec(`
 			UPDATE multipart_parts
 			SET content_md5 = $1, content_sha256 = $2, content_length = $3
-			WHERE upload_id = $4 AND part_number = $5
-		`, sqlMD5(contentMD5), sqlSHA256(sha256Bytes), contentLength, sqlUploadID(uid), partNumber)
+			WHERE multipart_upload_id = $4 AND part_number = $5
+		`, sqlMD5(contentMD5), sqlSHA256(sha256Bytes), contentLength, uid, partNumber)
 		if err != nil {
 			return fmt.Errorf("failed to update multipart part: %w", err)
 		}
+
 		n, err := res.RowsAffected()
 		if err != nil {
 			return err
 		} else if n == 0 {
-			return fmt.Errorf("no rows affected when updating multipart part")
+			return s3errs.ErrNoSuchUpload
 		}
 		return nil
 	})
